refactor: rename auth utility to crypto
