+++ backend/util/crypto.py
